server/requirements.txt
aiohttp>=3.11.11
fastapi>=0.96.0
openai>=1.59.3
pydantic>=1.10.19
python-dotenv>=1.0.1
python-multipart>=0.0.20
sqlalchemy>=2.0.15
uvicorn>=0.34.0
websockets>=14.1
langchain>=0.3.15
httpx>=0.26.0
aiofiles>=23.2.1
docker>=7.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
asyncpg>=0.29.0